Update "therock" target to build in a PyPA "manylinux" container (#438)
* Update "therock" target to build in a PyPA "manylinux" container

Change from gfx arch 4th argument to `build.sh` to a (gfx-specific)
python index URL argument to simplify upstreaming into TheRock.

Use a PyPA manylinux container as a base image for "therock" builds to
match TheRock's upstream builds and to maximize compatibility with as
many distributions as possible. The toolchains in those are specially
curated for that purpose. The images also come with all python
versions pre-installed.

Add a 5th argument to `build.sh` & `build_and_install.sh` so a custom
base image can be specified, if needed.

Add "torch" and "numpy" as dependence for the "uccl[rocm]" installs
build with TheRock (python packaged ROCm), as opposed to "rocm" which
is the system-installed ROCm target. UCCL's collective and transfer
modules need torch and it in turn needs numpy. TheRock project builds
torch and numpy python packages compatible with python packaged ROCm,
hence the dependence.

* Update Dockerfile.therock

Use the actual current TheRock base image. The PyPA image it is based on is in the comment above.

---------

Co-authored-by: Yang Zhou <yangzhou.rpc@gmail.com>