Add detection of stale builder image. (#556)
If a UCCL builder image exists check if it is older than the
Dockerfile to generate it. If yes, suggest removing the image and
pruning builder cache so it can be re-generated, but continue the
build after the warning and brief delay.

This can help detect spurious build issues if an older image cannot
support a new project or build feature.

Co-authored-by: Yang Zhou <yangzhou.rpc@gmail.com>